Refactor verifySignature for re-use
